requirements.txt
pandas
scikit-learn
sentence-transformers
faiss-cpu
wget
unzip
huggingface_hub[hf_xet]
google-generativeai
es_core_news_sm
spacy
matplotlib
seaborn
wordcloud
streamlit